Create skeleton of folders and move tcp defines
