+++ plugins/arch/CMakeLists.txt
@@ -1,3 +1,5 @@
 SET(SOURCE_GROUP ${SOURCE_GROUP}\\arch)
 
+add_subdirectory(posix)
+add_subdirectory(win32)
 add_subdirectory(arduino)
