+++ plugins/arch/win32/CMakeLists.txt
@@ -1,0 +1,9 @@
+SET(SOURCE_GROUP ${SOURCE_GROUP}\\win32)
+
+ua_add_architecture("win32")
+
+if("${UA_ARCHITECTURE}" STREQUAL "win32")
+
+ua_include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+
+endif()
