+++ tools/cmake/macros.cmake
@@ -3,3 +3,9 @@ FUNCTION(ua_add_architecture)
         set_property(GLOBAL APPEND PROPERTY UA_ARCHITECTURES ${ARG})
     ENDFOREACH(ARG)
 ENDFUNCTION(ua_add_architecture)
+
+FUNCTION(ua_include_directories)
+    FOREACH(ARG ${ARGV})
+        set_property(GLOBAL APPEND PROPERTY UA_INCLUDE_DIRECTORIES ${ARG})
+    ENDFOREACH(ARG)
+ENDFUNCTION(ua_include_directories)
